mallon

+++ src/main/java/com/mallonflowerz/almacen/facturasDeVentas/services/DetalleFacturaService.java
@@ -1,6 +1,7 @@
 package com.mallonflowerz.almacen.facturasDeVentas.services;
 
 import java.math.BigDecimal;
+import java.util.List;
 import java.util.Optional;
 import java.util.UUID;
 
@@ -9,7 +10,9 @@ import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 
 import com.mallonflowerz.almacen.facturasDeVentas.models.entity.DetalleFactura;
+import com.mallonflowerz.almacen.facturasDeVentas.models.entity.Factura;
 import com.mallonflowerz.almacen.facturasDeVentas.repositories.DetalleFacturaRepository;
+import com.mallonflowerz.almacen.facturasDeVentas.repositories.FacturaRepository;
 import com.mallonflowerz.almacen.productosYUsuarios.models.entity.Producto;
 import com.mallonflowerz.almacen.productosYUsuarios.repositories.ProductoRepository;
 
@@ -21,6 +24,7 @@ public class DetalleFacturaService {
 
     private final DetalleFacturaRepository dfRepository;
     private final ProductoRepository productoRepository;
+    private final FacturaRepository facturaRepository;
 
     public Page<DetalleFactura> listarDetalleFacturas(Pageable pageable) {
         return dfRepository.findAll(pageable);
@@ -51,7 +55,9 @@ public class DetalleFacturaService {
 
     public boolean eliminarDetalleFactura(UUID id) {
         Optional<DetalleFactura> detalleFacturaOp = dfRepository.findById(id);
-        if (detalleFacturaOp.isPresent()) {
+        List<Factura> facturas = facturaRepository.findAllByDetalles(detalleFacturaOp.get());
+        if (detalleFacturaOp.isPresent() && !facturas.isEmpty()) {
+            facturaRepository.deleteAll(facturas);
             dfRepository.delete(detalleFacturaOp.get());
             return true;
         }
